**New Commands!**

- Cleverbot
- Expand (expands short URLs)
- Level (Checks your level (works with point))
- Loadcommand (This allows you to load a new command without having to restart the bot)
- Lockdown (Locks a channel)
- Mute (Mutes a user (prevents them from talking))
- Permlevel (Check what permission level you have)
- Points (Check how many points you have)
- Reason (Allows a moderator to update a reason for a case)
- Remindme (Set reminders. Note: these will be reset when the bot restarts or the command is reloaded)
- Role (Allows moderators or above to give roles to members)
- Shorten (Allows you to shorten a long URL)
- Tags (Allows a moderator to store responses and a user to access these using `[prefix]tag tagname`)
- Weather (Now you can check the weather from Discord)

**New Music Commands!**

Music commands were added into the bot in this update:
- Np (Allows you to check what's playing)
- Pause (Pause the current song)
- Play (Plays a song)
- Queue (Check what's coming up)
- Resume (Continue a paused song)
- Skip (Allows you to skip a song)
- Stop (Stops the music, clears the queue, and leaves the channel)

*Please note that the music features are basic. I still need to add in a check for if there is no members in a voice channel that the music bot leaves that voice channel to save memory.*

**New Features!**

- Music Commands!
- Basic level system (You get points by sending messages)
- Case numbers

**Removed Feature**

- Removed the `blockConfigEval` option from the configuration file and removed this feature.

+++ events/message.js
@@ -1,7 +1,7 @@
 // The MESSAGE event runs anytime a message is received
 // Note that due to the binding of client to every event, every event
 // goes `client, other, args` when this function is run.
-const Discord = require('discord.js');
+//const Discord = require('discord.js');
 module.exports = (client, message) => { // eslint-disable-line complexity
 	// It's good practice to ignore other bots. This also makes your bot ignore itself
 	// and not get into a spam loop (we call that "botception").
@@ -9,15 +9,20 @@ module.exports = (client, message) => { // eslint-disable-line complexity
 	if (message.author.bot) {
 		return;
 	}
+
+	client.pointsMonitor(client, message);
+
 	/*if (message.channel.type === 'dm') {
 		return;
 	}*/
 
-	if (message.channel.type === 'dm') {
-		if (message.content.indexOf(client.config.defaultSettings.prefix) !== 0) {
-			return;
-		}
-	}
+	const settings = message.guild ? client.settings.get(message.guild.id) : client.config.defaultSettings;
+
+	//if (message.channel.type === 'dm') {
+	//if (message.content.indexOf(client.config.defaultSettings.prefix) !== 0) {
+	//return;
+	//}
+	//}
 
 
 	// Here we separate our "command" name, and our "arguments" for the command.
@@ -26,16 +31,16 @@ module.exports = (client, message) => { // eslint-disable-line complexity
 	// args = ["Is", "this", "the", "real", "life?"]
 	const args = message.content.split(/\s+/g);
 	var command;
-	if (message.channel.type !== 'dm') {
-		const gS = client.settings.get(message.guild.id); // TODO: Allow for per guild command disables
-		message.settings = gS;
-		command = args.shift().slice(gS.prefix.length)
-			.toLowerCase();
-	} else {
-		command = args.shift().slice(client.config.defaultSettings.prefix.length)
-			.toLowerCase();
-	}
 
+	//if (message.channel.type !== 'dm') {
+	//const gS = client.settings.get(message.guild.id); // TODO: Allow for per guild command disables
+	message.settings = settings;
+	command = args.shift().slice(settings.prefix.length)
+		.toLowerCase();
+	//} else {
+	//command = args.shift().slice(client.config.defaultSettings.prefix.length)
+	//.toLowerCase();
+	//}
 
 	// Get the user or member's permission level from the elevation
 	const level = client.permlevel(message);
@@ -54,22 +59,22 @@ module.exports = (client, message) => { // eslint-disable-line complexity
 	if (message.channel.type !== 'dm') {	// Grab the settings for this server from the enmap
 		const guildSettings = client.settings.get(message.guild.id); // TODO: Allow for per guild command disables
 
-		if (guildSettings.inviteFilterEnabled === 'true' || guildSettings.inviteFilterEnabled) {
-			if (message.content.match(/(discord\.(gg|me|io)|(discordapp\.com|discord\.com)\/invite).*/)) {
+		if (message.content.match(/(discord\.(gg|me|io)|(discordapp\.com|discord\.com)\/invite).*/) && guildSettings.inviteFilterEnabled === 'true') {
 
-				var msgInv = message.content.match(/discord\.gg\/[0-9A-Za-z-]+/);
-				//console.log(msgInv);
-				var dggInvCode = msgInv[0].replace(/discord\.gg\//, '');
-				//console.log(dggInvCode);
+			var msgInv = message.content.match(/discord\.gg\/[0-9A-Za-z-]+/);
+			//console.log(msgInv);
+			if (!msgInv) return;
+			var dggInvCode = msgInv[0].replace(/discord\.gg\//, '');
+			//console.log(dggInvCode);
 
-				var whitelist = guildSettings.inviteWhitelist;
-				//console.log(whitelist.includes(dggInvCode));
-				if (whitelist.includes(dggInvCode)) return;
-				message.delete();
-				message.reply('Invite links are not allowed');
-				//}
-				//}
+			var whitelist = guildSettings.inviteWhitelist;
+			//console.log(whitelist.includes(dggInvCode));
+			if (whitelist.includes(dggInvCode)) return;
+			if (level >= 2) { // Level 2 (moderator) or above is not affected by the invite link filter
+				return console.log(`${message.author.tag} (${message.author.id}) bypassed the invite link filter due to having a level of ${level}`);
 			}
+			message.delete();
+			message.reply('Invite links are not allowed');
 		}
 
 		if (guildSettings.swearFilter === 'true' && guildSettings.swearWords.some(word => message.content.includes(word))) {
@@ -88,78 +93,34 @@ module.exports = (client, message) => { // eslint-disable-line complexity
 			return;
 		}
 
+		if (client.talkedRecently.has(message.author.id)) {
+			return message.reply(`You need to wait ${parseInt(guildSettings.commandTimeout)}ms seconds between each command`);
+		}
 
-		if (guildSettings.logCommandUsage === 'true') {
+		if (level < 2) { // Level 2 (Moderator) and above do not have the cooldown
+			// Adds the user to the set so that they can't talk for 2.5 seconds
+			client.talkedRecently.add(message.author.id);
+			setTimeout(() => {
+				// Removes the user from the set after 2.5 seconds
+				client.talkedRecently.delete(message.author.id);
+			}, parseInt(guildSettings.commandTimeout));
+		}
+
+		if (guildSettings.logCommandUsage === 'true') { // Mod log channel command usage logs removed as it was giving issues with case numbers resetting if there was too many commands used and the last case went out of range of the fetch message limit
 			// If the command exists, **AND** the user has permission and it is not disabled, run it. Else, give error
 			if (cmd) {
 				if (level >= cmd.conf.permLevel) {
 					if (cmd.conf.enabled === true) {
-						const embed = new Discord.RichEmbed()
-							.setColor('RED')
-							.setTitle('Command Used')
-							.addField(`User`, `${message.author.tag} (${message.author.id})`, true)
-							.addField(`Command`, `${message.content}`, true)
-							.addField(`Channel`, `${message.channel.name} (${message.channel.id})`, true);
-						if (message.guild.channels.find('name', guildSettings.modLogChannel)) {
-							message.guild.channels.find('name', guildSettings.modLogChannel).send({ embed })
-								.catch((err) => {
-									console.log(err);
-								});
-						} else {
-							console.log(`Unable to send message to modLogChannel (${guildSettings.modLogChannel})`);
-						}
 						cmd.run(client, message, args, level);
 						console.log('log', `${message.guild.name}/#${message.channel.name} (${message.channel.id}):${message.author.username} (${message.author.id}) ran command ${message.content}`, 'CMD');
 					} else {
 						message.reply('This command is disabled');
-						const embed = new Discord.RichEmbed()
-							.setColor('RED')
-							.setTitle('Disabled Command Usage')
-							.addField(`User`, `${message.author.tag} (${message.author.id})`, true)
-							.addField(`Command`, `${message.content}`, true)
-							.addField(`Channel`, `${message.channel.name} (${message.channel.id})`, true);
-						if (message.guild.channels.find('name', guildSettings.modLogChannel)) {
-							message.guild.channels.find('name', guildSettings.modLogChannel).send({ embed })
-								.catch((err) => {
-									console.log(err);
-								});
-						} else {
-							console.log(`Unable to send message to modLogChannel (${guildSettings.modLogChannel})`);
-						}
 						client.log('log', `${message.guild.name}/#${message.channel.name} (${message.channel.id}):${message.author.username} (${message.author.id}) tried to run disabled command ${message.content}`, 'CMD');
 					}
 				} else {
-					const embed = new Discord.RichEmbed()
-						.setColor('RED')
-						.setTitle('No Permissions')
-						.addField(`User`, `${message.author.tag} (${message.author.id})`, true)
-						.addField(`Command`, `${message.content}`, true)
-						.addField(`Channel`, `${message.channel.name} (${message.channel.id})`, true);
-					if (message.guild.channels.find('name', guildSettings.modLogChannel)) {
-						message.guild.channels.find('name', guildSettings.modLogChannel).send({ embed })
-							.catch((err) => {
-								console.log(err);
-							});
-					} else {
-						console.log(`Unable to send message to modLogChannel (${guildSettings.modLogChannel})`);
-					}
 					client.log('log', `${message.guild.name}/#${message.channel.name} (${message.channel.id}):${message.author.username} (${message.author.id}) tried to run command ${message.content} without having the correct permission level`, 'CMD');
 				}
 			} else {
-				const embed = new Discord.RichEmbed()
-					.setColor('RED')
-					.setTitle('Non-existant Command')
-					.addField(`User`, `${message.author.tag} (${message.author.id})`, true)
-					.addField(`Command`, `${message.content}`, true)
-					.addField(`Channel`, `${message.channel.name} (${message.channel.id})`, true);
-				if (message.guild.channels.find('name', guildSettings.modLogChannel)) {
-					message.guild.channels.find('name', guildSettings.modLogChannel).send({ embed })
-						.catch((err) => {
-							console.log(err);
-						});
-				} else {
-					console.log(`Unable to send message to modLogChannel (${guildSettings.modLogChannel})`);
-				}
 				client.log('log', `${message.guild.name}/#${message.channel.name} (${message.channel.id}):${message.author.username} (${message.author.id}) tried to run non-existant command ${message.content}`, 'CMD');
 			}
 		} else {
